requirements.txt
pydantic~=2.9.2
requests~=2.32.3
python-dotenv~=1.0.1
Flask~=3.0.3